Rename from lds-gsim-spark to dapla-spark-plugin
And rename package to no.ssb.dapla.spark.plugin
