+++ pom.xml
@@ -4,11 +4,9 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
 
-    <groupId>no.ssb.lds</groupId>
-    <artifactId>lds-gsim-spark</artifactId>
+    <groupId>no.ssb.dapla.spark.plugin</groupId>
+    <artifactId>dapla-spark-plugin</artifactId>
     <version>0.3.0-SNAPSHOT</version>
-
-    <name>lds-gsim-spark</name>
 
     <properties>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
