+++ src/main/java/no/ssb/dapla/spark/plugin/DatasetDataPath.java
@@ -1,0 +1,4 @@
+package no.ssb.dapla.spark.plugin;
+
+public class DatasetDataPath {
+}
