+++ src/main/java/no/ssb/dapla/spark/plugin/DefaultSource.java
@@ -1,4 +1,4 @@
-package no.ssb.gsim.spark;
+package no.ssb.dapla.spark.plugin;
 
 public class DefaultSource extends GsimDatasource {
 }
