+++ src/main/java/no/ssb/dapla/spark/plugin/GsimRelation.java
@@ -1,4 +1,4 @@
-package no.ssb.gsim.spark;
+package no.ssb.dapla.spark.plugin;
 
 import org.apache.spark.rdd.RDD;
 import org.apache.spark.sql.Column;
