--- src/main/java/no/ssb/gsim/spark/DatasetDataPath.java
@@ -1,4 +1,0 @@
-package no.ssb.gsim.spark;
-
-public class DatasetDataPath {
-}
