+++ src/main/resources/META-INF/services/org.apache.spark.sql.sources.DataSourceRegister
@@ -1,1 +1,1 @@
-no.ssb.gsim.spark.GsimDatasource
+no.ssb.dapla.spark.plugin.GsimDatasource
