+++ src/test/java/no/ssb/dapla/spark/plugin/GsimRelationTest.java
@@ -1,4 +1,4 @@
-package no.ssb.gsim.spark;
+package no.ssb.dapla.spark.plugin;
 
 import org.junit.Test;
